Create Error component and http req error-handling functionality.
An Error component has been created consisting of a dialog modal, using
props to display an error alert and the error message, and it has a
button for the user to confirm and close the modal.

Within the "MealOffer" component, a new error State has been created and
it alternates between false or true depending on if the http request
is accomplished successfully or it fails. The try-catch error logic has
been built within the "fetchMeals" function; it updates the error State
to true when the http request fails, which triggers the render of the
"Error" modal component.