Implement "addItemToCart" functionality.
Updated the "shoppingCart" useState to a useReducer hook within the
"CartContextProvider" function.

Created "handleAddItemToCart" function within "CartContextProvider" as
a shoppingCart dispatch function to add items to the cart when the user
clicks the meal buttons from the "MealOffer" component. Added it to the
"cartCtx" variable so it is available to access from all components with
access to the context.

Created reducer outsider function "shoppingCartReducer" to handle the
shoppingCart State logic when dispatch functions are triggered. Added
a first if statement with the logic to add items when the user clicks
the buttons from the "MealOffer" component.

Updated the "MealOffer" component by accessing the context and
retrieving the "addItemToCart" to cart variable, which connects with
the "handleAddItemToCart" function from the context component.
Added also an "onClick" prop within the buttons and passed it the
"addItemToCart" and an object with the meal "id", "name", and "price"
as parameters, so that this data is available to use on the context
component.