Update "Header" and update and implement "Cart".
Header and Cart components have been updated to be linked by the
"forwardRef" hook, so that a ref within the Cart "dialog" element can
be accessed from Header component's "handleOpenCart" function to show
the modal in the UI.

Header also has been added the "handleCloseModal" function to close the
modal when the "Close" dialog button is clicked by the user.

Cart component has also been updated to show prices with two decimals at
most, and to render different content depending on if the user has
selected items or not.