Implement update/delete item functionalities within "Cart".
For the item update functionality:

--> Created "onClick" props within the -/+ button and passed a new
"updateCartItem" function which is coming from shopping-cart context.
"-" button uses a conditional statement within the prop so that it only
executes the updating function when the item quantity is over 1, so that
the item will have at least "1" of quantity at all times.

--> "updateCartItem" is passed and object with two values, the item id
and the type of action that the button executes (+ or -) to distinguish
them.

-->"updateCartItem" connects with the dispatch function
"handleUpdateCartItem" from the context, which in turn turn triggers the
"shoppingCartReducer" function. "shoppingCartReducer" detects the action
type "UPDATE_ITEM" from the dispatch function and increments or
diminishes the quantity of the specific item by one, depending on the
action of the button (+/-).

For the item delete functionality:

--> Used a trash "ion-icon" as if it were a button and added some css
to make it look good.

--> Created an "onClick" prop within the icon and set it to context
function "deleteCartItem", passing it only the item id as argument.

--> "deleteCartItem" triggers dispatch function "handleDeleteCartItem",
which trigges "shoppingCartReducer", which detects the "DELETE_ITEM"
type and filters through the state items array removing the item with
the same id as the one received.