Implement POST http request to store order data in the backend.
A new function "storeUserOrder" has been created within the "http.js"
file which sends a POST http request to the backend server with an
object containing an "items" key which expects an array of items and a
"customer" key which expects an object with the user form responses.

Within the "CheckOutForm" component, a "handleSubmit" function has been
created to trigger to an "onSubmission" prop within the form. When the
user submits the form, the prop's default behaviour is prevended and
instead the form data is gathered and then function "storeUserOrder" is
executed passing the "items" context array as the first argument, and
the "userOrder" object containing user form answers as the second, so
that this information is stored in the backend "orders.json" file.

Logic to discriminate fetching errors has been started but it must be
finished as a following step.