Add "OrderSubmitted" modal component + small fixes
"OrderSubmitted" modal component is created and shown within the header
section whenever the user submits the items order correctly from the
"CheckOutForm" modal. It displays a message informing that the order
has been processed.

"CheckOutForm's" component "handleSubmit" function has been tweaked to
accomplish:
--> closing the "checkOutForm" modal through the "OrderSubmitted()"
function which is passed on to "header" component as a prop, triggering
the "handleCloseModal" function.
--> Resetting the cart context items array by executing new context
function "clearCartItems()".

Cart context has been updated with the addition of function
"clearCartItems". It resets the State items array whenever it is called.
Also, function "calculateTotalPrice" has been tweaked to return the
total price with fixed decimals of two.

"Header" component imports now the "OrderSubmitted" modal and shows it
or closes it through "handleCloseModal" function.